pyproject.toml
[tool.poetry]
name = "messenger-mirror"
version = "0.1.0"
description = "Send email notification when receiving Facebook message."
authors = ["Radon Rosborough <radon.neon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^3.141.0"
chromedriver-py = "^94.0.4606"
python-dotenv = "^0.19.0"
persist-queue = "^0.7.0"
requests = "^2.26.0"
sendgrid = "^6.8.2"
Flask = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
